// src/lib/types/satellite.ts
export enum OrbitType {
  /** 低地球轨道 (<2000km) */
  LEO = 'LEO',
  /** 中地球轨道 (2000-35786km) */
  MEO = 'MEO',
  /** 地球同步轨道 (~35786km) */
  GEO = 'GEO',
  /** 高椭圆轨道 */
  HEO = 'HEO'
}